+++ tests/TestCase/DatabaseSuite.php
@@ -72,11 +72,11 @@ class DatabaseSuite extends TestSuite
         $permutations = [
             'Identifier Quoting' => function () {
                 ConnectionManager::get('test')
-                                 ->driver()
-                                 ->autoQuoting(true);
+                                 ->getDriver()
+                                 ->enableAutoQuoting();
 //             },
 //             'No identifier quoting' => function () {
-//                 ConnectionManager::get('test')->driver()->autoQuoting(false);
+//                 ConnectionManager::get('test')->getDriver()->enableAutoQuoting(false);
             }
         ];